Enhancement: Local development with Docker
This commit enables local development of the website with
Docker.

Note, I had to add a leading "/" at the `baseurl` in the
`_config.yml`, otherwise it would not work locally.

I'm assuming this will continue to work at the
current GitHub pages setup, but that should be tested,
and reverted if needed.

I also added a `.gitignore` and a `Gemfile` that should
also help with a native ruby + jekyll local development
environment.

This is a stepwise increment, I would like to have the
possibility to test the website locally also with a
local development version of the remote theme
`InnerSourceCommons/working-group-roles`.

Note that I copied the `Gemfile`, `docker-compose.yml`
and `Dockerfile.local` from an internal SAP repository
and adapted it. There might be some superfluous content.